factored out the code related to relationship out from Application.java to Relationships.java, also added relationships page
