+++ app/controllers/Application.java
@@ -14,10 +14,10 @@ public class Application extends OBController {
     render(user);
   }
 
-  public static void news(Long id) {
-    User user = id == null ? user() : (User) User.findById(id);
-    render(user);
-  }
+	public static void news(Long id) {
+		User user = id == null ? user() : (User) User.findById(id);
+		render(user);
+	}
 
   public static void account() {
     User user = user();
@@ -28,75 +28,6 @@ public class Application extends OBController {
     return val != null && val.length() > 0;
   }
 
-  /** Request to be friends with a given user, changing appropriate Relationship flags where necessary.
-   *
-   * @param id the user to request friendship with
-   */
-  public static void requestFriends(Long id) {
-    User user = user();
-    User other = User.findById(id);
-    Relationship r1 = Relationship.find("SELECT r FROM Relationship r where r.from = ? AND r.to = ?", user, other).first();
-    Relationship r2 = Relationship.find("SELECT r FROM Relationship r where r.to = ? AND r.from = ?", user, other).first();
-
-    if (r1 == null) {
-      r1 = new Relationship(user, other, true);
-    }
-
-    if (r2 != null) {
-
-      // If the other user has already requested, this request should make them friends
-      if (r2.requested) {
-        r1.requested = false;
-        r2.requested = false;
-        r1.accepted = true;
-        r2.accepted = true;
-        r1.save();
-        r2.save();
-      } else if (r2.accepted){
-        news(id);
-        return;
-      } else {
-        r1.requested = true;
-      }
-    }
-    // If the user has already made a request for friendship, do nothing
-    else if (r1 != null) {
-      if (r1.requested) {
-        news(id);
-        return;
-      } else {
-        r1.requested = true;
-        r1.save();
-        news(id);
-        return;
-      }
-    }
-    r1.save();
-    news(id);
-  }
-
-  /** Attempt to end a relationship with a user, changing appropriate Relationship flags where necessary
-   *
-   * @param id the user to remove
-   */
-  public static void removeFriends(Long id) {
-    User user = user();
-    User other = User.findById(id);
-    Relationship r1 = Relationship.find("SELECT r FROM Relationship r where r.from = ? AND r.to = ?", user, other).first();
-    Relationship r2 = Relationship.find("SELECT r FROM Relationship r where r.to = ? AND r.from = ?", user, other).first();
-
-    if (r1 != null) {
-      r1.requested = false;
-      r1.accepted = false;
-      r1.save();
-    }
-    if (r2 != null) {
-      r2.accepted = false;
-      r2.requested = false;
-      r2.save();
-    }
-    news(id);
-  }
 
   public static void account_save(User update, String old_password) {
     User currentUser = user();
