+++ app/views/Application/news.html
@@ -11,7 +11,7 @@
     <a href="@{Application.about()}">about</a>
     <br>
     <a href="@{Checkin.checkin()}">Checkin</a>
-    <br>friends:<br>
+    <br><a href="@{Relationships.relationships(user.id)}">friends:</a><br>
      #{list items:user.confirmedFriends(), as:'friendship'}
      <a href="@{Application.news(friendship.to.id)}">${ friendship.to.name }</a><br>
      #{/list}
@@ -19,7 +19,7 @@
   <td>
      <h1>${user.name}</h1> 
      #{if user.checkFriendship(user.id) == 'Request Friendship'}
-	     <a id="friend-button" href="@{Application.requestFriends(user.id)}">
+	     <a id="friend-button" href="@{Relationships.requestFriends(user.id)}">
 		  	${user.checkFriendship(user.id)}
 		     </a>
 	 #{/if}
@@ -28,7 +28,7 @@
 	 #{/elseif}
 	 #{elseif user.checkFriendship(user.id) == 'Friends'}
 	 	Friends :) 
-	 	<a id="friend-button" href="@{Application.removeFriends(user.id)}">
+	 	<a id="friend-button" href="@{Relationships.removeFriends(user.id)}">
 		  	Remove friend
 		</a>
 	#{/elseif}
