+++ conf/routes
@@ -43,7 +43,8 @@ GET     /users/{<[0-9]+>ownerId}/photos/?       Photos.photos
 #POST    /status/{<[0-9]+>statusId}/comments/?   #Add a comment to a status
 #GET     /users/{<[0-9]+>userId}/comments/?      #get all comments by a user
 
-
+#Relationships
+GET 	/relationships							Relationships.relationships
 
 # Checkin
 GET     /checkin                                Checkin.checkin
